# apps/api/docs/openapi.yaml
components:
    schemas:
        handler.achievementDTO:
            properties:
                achieved_at:
                    nullable: true
                    type: string
                community_id:
                    nullable: true
                    type: string
                description:
                    nullable: true
                    type: string
                icon_or_badge_url:
                    nullable: true
                    type: string
                id:
                    nullable: true
                    type: string
                member_id:
                    nullable: true
                    type: string
                title:
                    nullable: true
                    type: string
            type: object
...
        handler.upsertAchievementRequest:
            properties:
                achieved_at:
                    type: string
                description:
                    nullable: true
                    type: string
                icon_or_badge_url:
                    nullable: true
                    type: string
                member_id:
                    nullable: true
                    type: string
                title:
                    type: string
            required:
                - achieved_at
                - title
            type: object
...
        response.Envelope:
            properties:
                data:
                    nullable: true
                error:
                    $ref: '#/components/schemas/response.ErrorBody'
                meta:
                    $ref: '#/components/schemas/response.Meta'
                success:
                    nullable: true
                    type: boolean
            type: object
        response.ErrorBody:
            nullable: true
            properties:
                code:
                    $ref: '#/components/schemas/response.ErrorCode'
                fields:
                    additionalProperties:
                        type: string
                    nullable: true
                    type: object
                message:
                    nullable: true
                    type: string
            type: object
        response.ErrorCode:
            enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMITED
                - INTERNAL_ERROR
            nullable: true
            type: string
            x-enum-varnames:
                - ErrValidation
                - ErrUnauthorized
                - ErrForbidden
                - ErrNotFound
                - ErrConflict
                - ErrRateLimited
                - ErrInternal
        response.Meta:
            nullable: true
            properties:
                page:
                    nullable: true
                    type: integer
                per_page:
                    nullable: true
                    type: integer
                total:
                    nullable: true
                    type: integer
            type: object
    securitySchemes:
        BearerAuth:
            description: Clerk session JWT, sent as "Bearer <token>".
            in: header
            name: Authorization
            type: apiKey
info:
    contact: {}
    description: Backend API for the Iridescent community hub (events, memberships, announcements, gallery, achievements).
    title: Iridescent Hub API
    version: "1.0"
openapi: 3.0.3
paths:
    /achievements:
        get:
            parameters:
                - description: Page number
                  in: query
                  name: page
                  schema:
                    default: 1
                    type: integer
                - description: Items per page
                  in: query
                  name: per_page
                  schema:
                    default: 20
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/response.Envelope'
                                    - properties:
                                        data:
                                            items:
                                                $ref: '#/components/schemas/handler.achievementDTO'
                                            type: array
                                      type: object
                    description: OK
            summary: List achievements
            tags:
                - achievements
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/handler.upsertAchievementRequest'
                description: Achievement
                required: true
                x-originalParamName: body
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/response.Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/handler.achievementDTO'
                                      type: object
                    description: Created
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.Envelope'
                    description: Unprocessable Entity
            security:
                - BearerAuth: []
            summary: Create an achievement
            tags:
                - achievements
    /achievements/{id}:
        delete:
            parameters:
                - description: Achievement ID
                  in: path
                  name: id
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.Envelope'
                    description: Not Found
            security:
                - BearerAuth: []
            summary: Delete an achievement
            tags:
                - achievements
        patch:
            parameters:
                - description: Achievement ID
                  in: path
                  name: id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/handler.upsertAchievementRequest'
                description: Achievement
                required: true
                x-originalParamName: body
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/response.Envelope'
                                    - properties:
                                        data:
                                            $ref: '#/components/schemas/handler.achievementDTO'
                                      type: object
                    description: OK
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.Envelope'
                    description: Unprocessable Entity
            security:
                - BearerAuth: []
            summary: Update an achievement
            tags:
                - achievements
    /admin/dashboard/summary:
        get:
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.Envelope'
                    description: OK
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.Envelope'
                    description: Forbidden
            security:
                - BearerAuth: []
            summary: Admin dashboard summary
            tags:
                - admin
...
servers:
    - url: /api/v1